+++ CHANGES.md
@@ -15,6 +15,7 @@ Changes
 - Add check UseValueCompare
 - ScopeBuilderVisitor now also tracks usage
 - Fix ScopeBuilderVisitor uncorrectly marking ScopeEntry as GLOBAL
+- Fix locating magik-lint.properties in products
 
 
 0.3.2 (2019-10-29)